% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dar.comparisons}
\alias{dar.comparisons}
\title{Comparative Judgement on Deprivation in Dar es Salaam, Tanzania}
\format{
A csv file containing 76408 rows and 2 columns. Each row corresponds to a judgement made by a single judge. The column shows which of the pair of subwards was judged to be poorest and richest.
}
\usage{
system.file("extdata", "dar.csv", package = "BTUN")
}
\description{
A comparative judgement dataset on deprivation in subwards in Dar es Salaam, Tanzania.
Citizens were shown pairs of subwards at random and asked which was more deprived.
If they said they were equal, one of the pair was chosen at random to be more deprived.
The data was collected in August 2018.
}
\keyword{datasets}
