% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dar.shapefiles}
\alias{dar.shapefiles}
\title{Shape files for the subwards in Dar es Salaam, Tanzania}
\format{
A .shp object
}
\usage{
dar.shapefiles
}
\description{
Polygons for the 452 subwards in Dar es Salaam, Tanzania
}
\keyword{datasets}
