% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{english.laa.adj.matrix}
\alias{english.laa.adj.matrix}
\title{Adjacency matrix for the Local Authority Areas in England}
\format{
A 4317x317 matrix, where a_{ij} = 1 if LAAs i and j are neighbours and 0 otherwise. The adjacency matrix is based on areas which share administrative borders. An edge has been added between the Isles of Scilly and Cornwall. 6 edges have been added between the Isle of Wight and LAAs in south Hampshire.
}
\usage{
english.laa.adj.matrix
}
\description{
Adjacency matrix for the Local Authority Areas in England
}
\keyword{datasets}
