% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{english.laa.deprivation}
\alias{english.laa.deprivation}
\title{Indicies of Multiple Deprivation for Local Authority Areas in England}
\format{
A .shp object
}
\source{
\rul{https://www.gov.uk/government/statistics/english-indices-of-deprivation-2019}
}
\usage{
english.laa.deprivation
}
\description{
A data frame containing the average index of multiple deprivation for each LAA in England.
}
\keyword{datasets}
