% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{english.laa.shapefiles}
\alias{english.laa.shapefiles}
\title{Shape files for Local Authority Areas in England}
\format{
A .shp object
}
\source{
\rul{https://geoportal.statistics.gov.uk/datasets/ae90afc385c04d869bc8cf8890bd1bcd_1}
}
\usage{
english.laa.shapefiles
}
\description{
Polygons for the 317 local authority areas in england. Names and coordinates are also included.
}
\keyword{datasets}
