\name{forcedmarriage.comparisons}
\alias{forcedmarriage.comparisons}
\docType{data}
\title{Comparative Judgment on Forced Marriage in Nottinghamshire, UK}
\format{
A csv file containing 1846 rows and 4 columns. Each row corresponds to a judgement made by a single judge. Columns 3 and 4 shows which of the pair of wards was judged to have relatively higher and low forced marriage risk level, column 1 shows which judge the comparison belong to, and column 2 shows what time they made the decision.
}
\source{
The data was collected using support from the Engineering and Physical Sciences Research Council [grant reference EP/R513283/1], the Economic and Social Sciences Research Council [ES/V015370/1] and the Research England Policy Support Fund. The data was collected following ethical approval from the University of Nottingham School of Politics and International Relations ethics committee.
}
\usage{
forcedmarriage.comparisons
}
\description{
A comparative judgment data set for risk of forced marriage at ward level in Nottinghamshire. There are 12 judges and 76 wards. Each comparison was made from 2 randomly selected wards.
The time of each comparision is also included.
}
\keyword{datasets}
