% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{loglike_function}
\alias{loglike_function}
\title{Compute the value of the loglikelihood function}
\usage{
loglike_function(x, win.matrix)
}
\arguments{
\item{x}{The level of deprivation of the areas on an exponential scale}

\item{win.matrix}{A matrix, where w_ij give the number of times area i beat j}
}
\value{
The value of of the loglikelihood function
}
\description{
This function computes the value of the binomial loglikelihood function
}
