% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{male.mean.deprivation}
\alias{male.mean.deprivation}
\title{The mean level of deprivation for subwards in Dar es Salaam as perceived by men}
\format{
An vector of 452 elements, one for each subward
}
\usage{
male.mean.deprivation
}
\description{
This data is used in the vignette
}
\keyword{datasets}
