% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcmc.output}
\alias{mcmc.output}
\title{MCMC Output for Dar es Salaam Vignette}
\format{
A list containing the value of f and alpha at each iteration of the MCMC output
}
\usage{
mcmc.output
}
\description{
MCMC Output for Dar es Salaam Vignette
}
\keyword{datasets}
