% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mean.deprivation}
\alias{mean.deprivation}
\title{The Mean Level of Deprivation for Subwards in Dar es Salaam}
\format{
An vector
}
\usage{
mean.deprivation
}
\description{
This data is used in the vignette
}
\keyword{datasets}
