% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{mvnorm_chol}
\alias{mvnorm_chol}
\title{Draw a sample from a multivariate normal diastirbution}
\usage{
mvnorm_chol(mu, chol)
}
\arguments{
\item{mu}{The mean vector}

\item{chol}{The cholesky decomposition of the covariance matrix Sigma}
}
\value{
a vector containing a sample from the distribution
}
\description{
This draws a sample from a multivariate normal distribution with mean vector mu and covariance matirx Sigma, using the Cholesky decomposition method.
}
