% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{quality_ratio}
\alias{quality_ratio}
\title{Compute the quality ratio on a logit scale}
\usage{
quality_ratio(x, y)
}
\arguments{
\item{x}{The level of deprivation of the winning area on the exponential scale}

\item{y}{The level of deprivation of the losing area on the exponential scale}
}
\value{
logit(pi_ij)
}
\description{
This function compute the logit(pi_ij) = lambda_i - log(exp(lambda_i) + exp(lambda_j))
}
