% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registered.network.covariance.function.R
\name{registered.network.covariance.function}
\alias{registered.network.covariance.function}
\title{Construct a registered covariance matrix from the coordiantes of the nodes}
\usage{
registered.network.covariance.function(
  coordinates,
  plane.boundary,
  type,
  hyperparameters,
  linear.combination,
  linear.constraint = 0
)
}
\arguments{
\item{type}{The type of covariance function used. One of "sqexp", "ratquad" or "matern". Note: only matern with nu = 5/2 is supported.}

\item{linear.combination}{A matrix which defines the linear combination of (lambda_1, ..., lambda_N)^T.}

\item{linear.constraint}{The value the linear constraint takes. Defaults to 0.}

\item{coordiantes}{An Nx2 matrix containing the Euclidean coordinates of the nodes.}

\item{hyperaparameters}{A vector containing the covariance function hyperparameters. For the squared exponential and matern, the vector should contain the variance and length scale, for the rational quadratic, the vector should contain the variance, lenght scale and scaling parameters}
}
\value{
The mean vector and covariance matrix
}
\description{
This function constructs a covariance function from the Euclidean coordinates of the nodes. It constructs a voronoi grpah to partition the space and places edges betweeen areas which are adjacent. The covairance function may be suqred exponential, rational quadratic or Matern
}
