% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registered.covariance.function.R
\name{registered_covariance_function}
\alias{registered_covariance_function}
\title{Construct a registered covariance matrix from the Euclidean coordinates of the objects}
\usage{
registered_covariance_function(
  coordinates,
  type,
  hyperparameters,
  linear.combination,
  linear.constraint = 0
)
}
\arguments{
\item{coordinates}{An Nx2 matrix containing the Euclidean coordinates of the nodes.}

\item{type}{The type of covariance function used. One of "sqexp", "ratquad" or "matern". Note: only matern with nu = 5/2 is supported.}

\item{hyperparameters}{A vector containing the covariance function hyperparameters. For the squared exponential and matern, the vector should contain the variance and length scale, for the rational quadratic, the vector should contain the variance, lenght scale and scaling parameters}

\item{linear.combination}{A matrix which defines the linear combination of (lambda_1, ..., lambda_N)^T.}

\item{linear.constraint}{The value the linear constraint takes. Defaults to 0.}
}
\value{
The mean vector and covariance matrix
}
\description{
This function constructs a covariance function from the Euclidean coordinates of the objects. The covairance function may be squared exponential, rational quadratic or Matern.
}
