% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run the BSBT MCMC algorithm}
\usage{
run_mcmc(n.iter, delta, k.mean, k.chol, win.matrix, f.initial, alpha = FALSE)
}
\arguments{
\item{n.iter}{The number of iterations to be run}

\item{delta}{The underrlaxed tuning parameter must be in (0, 1)}

\item{k.mean}{The prior mean vector}

\item{k.chol}{The cholesky decomposition of the prior covariance matrix}

\item{win.matrix}{A matrix, where w_ij give the number of times area i beat j}

\item{f.initial}{A vector of the initial estimate for f}

\item{alpha}{A boolean if inference for alpha should be carried out}
}
\value{
A list of MCMC output
\itemize{
\item f.matrix - A matrix containing the each iteration of f
\item alpha.sq - A vector containing the iterations of alpha^2
\item accpetance.rate - The acceptance rate for f
\item time.taken - Time taken to run the MCMC algorithm in seconds
}
}
\description{
This function runs the BSBT MCMC algorithm to estimate the deprivation parameters. In this version, the judges are assumed to act homogeneously. This algorithm estimates the deprivation in each area and the prior distribution variance parameter. For data with two types of judges, see \code{\link{run_gender_mcmc}}.
}
\examples{

n.iter <- 10
delta <- 0.1
k.mean <- c(0, 0, 0)
k.chol <- diag(3) #decomposed covariance matrix
comparisons <- data.frame("winner" = c(1, 3, 2, 2), "loser" = c(3, 1, 1, 3))
win.matrix <- comparisons_to_matrix(3, comparisons) #construct covariance matrix
f.initial <- c(0, 0, 0) #intial estimates for lamabda_1, lambda_2, lambda_3

mcmc.output <- run_mcmc(n.iter, delta, k.mean, k.chol, win.matrix, f.initial)


}
