% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{run_mcmc_with_ordering}
\alias{run_mcmc_with_ordering}
\title{Run the BSBT MCMC algorithm with ordering constraints}
\usage{
run_mcmc_with_ordering(
  n.iter,
  delta,
  k.mean,
  k.chol,
  win.matrix,
  f.initial,
  S,
  alpha = FALSE
)
}
\arguments{
\item{n.iter}{The number of iterations to be run}

\item{delta}{The underrlaxed tuning parameter must be in (0, 1)}

\item{k.mean}{The prior mean vector}

\item{k.chol}{The cholesky decomposition of the prior covariance matrix}

\item{win.matrix}{A matrix, where w_ij give the number of times area i beat j}

\item{f.initial}{A vector of the intial esitmate for f}

\item{S}{A list of ordering constraints. There are four elements in each set, the label of the two areas, the value of the constaint, and the confidence parameter.}

\item{alpha}{A boolean if inference for alpha should be carried out}
}
\value{
A list of MCMC output
\itemize{
\item f.matrix - A matrix containing the each iteration of f
\item alpha.sq - A vector containing the iterations of alpha^2
\item accpetance.rate - The acceptance rate for f
\item time.taken - Time tkane to run the MCMC algorithm in seconds
}
}
\description{
This function runs the BSBT mcmc algorithm with ordering constraints. The constraints are
included using a list of sets.
}
\examples{

n.iter <- 10
delta <- 0.1
k.mean <- c(0, 0, 0)
k.chol <- diag(3)
comparisons <- data.frame("winner" = c(1, 3, 2, 2), "loser" = c(3, 1, 1, 3))
win.matrix <- comparisons_to_matrix(3, comparisons)
f.initial <- c(0, 0, 0)
S <- list()
S[[1]] <- c(1, 3, -1, 3)
S[[2]] <- c(1, 2, -1, 3)
mcmc.output <- run_mcmc_with_ordering(n.iter, delta, k.mean, k.chol, win.matrix, f.initial, S)


}
