% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_contests.R
\name{simulate_comparisons}
\alias{simulate_comparisons}
\title{Simulate contests from the Bradley--Terry Model}
\usage{
simulate_comparisons(n.contests, true.quality, sigma.obs)
}
\arguments{
\item{n.contests}{The number of contests to be carried out}

\item{true.quality}{A vector with the level of deprivation in each area on the log scale.}

\item{sigma.obs}{Standard deviation for the noise to be added to the level of deprivation in each subward. If 0, no noise is used.}
}
\value{
A list containing a data.frame with each par-wise contest and the outcome, and a win matrix where the i,j^th element is the number of times i beat j
}
\description{
This function simulates pair-wise contests according to the Bradley--Terry model
}
\examples{

example.deprivation <- -2:2 #True level of deprivation in each area
example.comparisons <- simulate_comparisons(10, example.deprivation, 0) #generate comparisons

}
