% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dar.comparisons}
\alias{dar.comparisons}
\title{Comparative Judgement on Deprivation in Dar es Salaam, Tanzania}
\format{
A csv file containing 76408 rows and 3 columns. Each row corresponds to a judgement made by a single judge. Columns 2 and 3 shows which of the pair of subwards was judged to be poorest and richest, and column 3 shows the genderr of the judge.
}
\source{
This dataset was collected by Madeleine Ellis, James Goulding, Bertrand Perrat, Gavin Smith and Gregor Engelmann.We gratefully acknowledge the Rights Lab at the University of Nottingham for supporting funding for the comprehensive ground truth survey. We also acknowledge HumanitarianStreet Mapping Team (HOT) for providing a team of experts in data collectionto facilitate the surveys. This work was also supported by the EPSRC HorizonCentre for Doctoral Training - My Life in Data (EP/L015463/1) and EPSRC grant Neodemographics (EP/L021080/1).
}
\usage{
dar.comparisons
}
\description{
A comparative judgement dataset on deprivation in subwards in Dar es Salaam, Tanzania.
Citizens were shown pairs of subwards at random and asked which was more deprived.
If they said they were equal, one of the pair was chosen at random to be more deprived.
The data was collected in August 2018. The gender of each judge is also included.
}
\keyword{datasets}
