% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{loglike_function}
\alias{loglike_function}
\title{Compute the loglikelihood function}
\usage{
loglike_function(x, win.matrix)
}
\arguments{
\item{x}{The level of deprivation of the areas on an exponential scale}

\item{win.matrix}{A matrix, where w_ij give the number of times area i beat j}
}
\value{
The value of of the loglikelihood function
}
\description{
This function computes the BSBT model loglikelihood function. It requires the deprivation levels and the win matrix.
}
\examples{

win.matrix <- matrix(c(0, 3, 2, 1, 0, 1, 1, 3, 0), 3, 3) #construct win matrix
lambda     <- c(3, 1, 2)

l <- loglike_function(lambda, win.matrix)

}
\keyword{interal}
