% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{quality_ratio}
\alias{quality_ratio}
\title{Compute the logit of the ratio of qualities}
\usage{
quality_ratio(x, y)
}
\arguments{
\item{x}{The level of deprivation of the winning area on the exponential scale}

\item{y}{The level of deprivation of the losing area on the exponential scale}
}
\value{
logit(pi_ij)
}
\description{
This function computes the probability i beats j on the logit scale. It is given by logit(pi_ij) = lambda_i - log(exp(lambda_i) + exp(lambda_j))
}
\examples{

#compare areas with levels -2 and 2

qr <- quality_ratio(exp(-2), exp(2))

}
\keyword{internal}
