% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{run_gender_mcmc}
\alias{run_gender_mcmc}
\title{Run the BTUN with Gender Effect MCMC algorithm}
\usage{
run_gender_mcmc(
  n.iter,
  delta,
  k.mean,
  k.chol,
  male.win.matrix,
  female.win.matrix,
  f.initial,
  g.initial
)
}
\arguments{
\item{n.iter}{The number of iterations to be run}

\item{delta}{The underrlaxed tuning parameter must be in (0, 1)}

\item{k.mean}{The GP prior mean vector}

\item{k.chol}{The cholesky decomposition of the GP prior covariance matrix, alpha must be set to 1 when constructing ths}

\item{male.win.matrix}{A matrix, where w_ij give the number of times area i beat j when judged by men}

\item{female.win.matrix}{A matrix, where w_ij give the number of times area i beat j when judged by women}

\item{f.initial}{A vector of the intial esitmate for f, the male function}

\item{g.initial}{A vector of the intial esitmate for g, the discrepancy functon}
}
\value{
A list of MCMC output
\itemize{
\item f.matrix - A matrix containing the each iteration of f
\item g.matrix - A matrix containing the each iteration of g
\item alpha.sq - A matrix containing the iterations of alpha^2
\item accpetance.rate - The acceptance rate for f and g
\item time.taken - Time tkane to run the MCMC algorithm in seconds
}
}
\description{
This function runs the BTUN with Gender Effect MCMC algorithm
}
\examples{

n.iter <- 10
delta <- 0.1
k.mean <- c(0, 0, 0)
k.chol <- diag(3)
men.comparisons <- data.frame("winner" = c(1, 3, 2, 2), "loser" = c(3, 1, 1, 3))
women.comparisons <- data.frame("winner" = c(1, 2, 1, 2), "loser" = c(3, 1, 3, 3))
men.win.matrix <- comparisons_to_matrix(3, men.comparisons)
women.win.matrix <- comparisons_to_matrix(3, women.comparisons)
f.initial <- c(0, 0, 0)
g.initial <- c(0, 0, 0)

mcmc.output <- run_gender_mcmc(n.iter, delta, k.mean, k.chol, men.win.matrix,
    women.win.matrix, f.initial, g.initial)

}
