% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run the BTUN MCMC algorithm}
\usage{
run_mcmc(n.iter, delta, k.mean, k.chol, win.matrix, f.initial, alpha = FALSE)
}
\arguments{
\item{n.iter}{The number of iterations to be run}

\item{delta}{The underrlaxed tuning parameter must be in (0, 1)}

\item{k.mean}{The GP prior mean vector}

\item{k.chol}{The cholesky decomposition of the GP prior covariance matrix}

\item{win.matrix}{A matrix, where w_ij give the number of times area i beat j}

\item{f.initial}{A vector of the intial esitmate for f}

\item{alpha}{A boolean if inference for alpha should be carried out}
}
\value{
A list of MCMC output
\itemize{
\item f.matrix - A matrix containing the each iteration of f
\item alpha.sq - A vector containing the iterations of alpha^2
\item accpetance.rate - The acceptance rate for f
\item time.taken - Time tkane to run the MCMC algorithm in seconds
}
}
\description{
This function runs the BTUN mcmc algorithm
}
\examples{

n.iter <- 10
delta <- 0.1
k.mean <- c(0, 0, 0)
k.chol <- diag(3)
comparisons <- data.frame("winner" = c(1, 3, 2, 2), "loser" = c(3, 1, 1, 3))
win.matrix <- comparisons_to_matrix(3, comparisons)
f.initial <- c(0, 0, 0)

mcmc.output <- run_mcmc(n.iter, delta, k.mean, k.chol, win.matrix, f.initial)


}
